Criado o projeto de testes e recriado a migração incial

+++ src/Volvo.Services.Trucks.Domain/Commands/Trucks/Delete/DeleteTruckCommandHandler.cs
@@ -23,6 +23,14 @@ namespace Volvo.Services.Trucks.Domain.Commands.Trucks.Delete
             CancellationToken cancellationToken
         )
         {
+            var truck = await _repository.GetSingleAsync(x => x.Id == request.Id,false);
+
+            if(truck == default)
+            {
+                AddNotFoundNotification();
+                return default;
+            }
+
             await _repository.DeleteAsync(request.Id);
             await _unitOfWork.SaveChangesAsync();
             return Unit.Value;
